Get UTXO DB unit tests passing
