+++ testkit/src/main/scala/org/bitcoins/testkit/wallet/WalletTestUtil.scala
@@ -14,11 +14,14 @@ import scodec.bits.HexStringSyntax
 import org.bitcoins.core.hd._
 import org.bitcoins.core.protocol.script.ScriptWitness
 import org.bitcoins.core.protocol.script.P2WPKHWitnessV0
-import org.bitcoins.core.protocol.transaction.TransactionOutput
-import org.bitcoins.core.protocol.transaction.TransactionOutPoint
-import org.bitcoins.wallet.models.NativeV0UTXOSpendingInfoDb
-import org.bitcoins.core.currency._
-import org.bitcoins.wallet.models.LegacyUTXOSpendingInfoDb
+import org.bitcoins.core.protocol.transaction.Transaction
+import org.bitcoins.testkit.core.gen.TransactionGenerators
+import scala.concurrent.Future
+import org.bitcoins.wallet.models.IncomingTransaction
+import org.bitcoins.wallet.models.AddressDb
+import org.bitcoins.wallet.models.AddressDbHelper
+import org.bitcoins.testkit.fixtures.WalletDAOs
+import scala.concurrent.ExecutionContext
 
 object WalletTestUtil {
 
@@ -70,30 +73,62 @@ object WalletTestUtil {
   lazy val sampleSPK: ScriptPubKey =
     ScriptPubKey.fromAsmBytes(hex"001401b2ac67587e4b603bb3ad709a8102c30113892d")
 
-  lazy val sampleSegwitUtxo: NativeV0UTXOSpendingInfoDb = {
-    val outpoint =
-      TransactionOutPoint(WalletTestUtil.sampleTxid, WalletTestUtil.sampleVout)
-    val output = TransactionOutput(1.bitcoin, WalletTestUtil.sampleSPK)
-    val scriptWitness = WalletTestUtil.sampleScriptWitness
-    val privkeyPath = WalletTestUtil.sampleSegwitPath
-    NativeV0UTXOSpendingInfoDb(id = None,
-                               outPoint = outpoint,
-                               output = output,
-                               privKeyPath = privkeyPath,
-                               scriptWitness = scriptWitness,
-                               incomingTxId = None)
-  }
-
-  lazy val sampleLegacyUtxo = {
-    val outpoint =
-      TransactionOutPoint(WalletTestUtil.sampleTxid, WalletTestUtil.sampleVout)
-    val output = TransactionOutput(1.bitcoin, WalletTestUtil.sampleSPK)
-    val privKeyPath = WalletTestUtil.sampleLegacyPath
-    LegacyUTXOSpendingInfoDb(id = None,
-                             outPoint = outpoint,
-                             output = output,
-                             privKeyPath = privKeyPath,
-                             incomingTxId = None)
-  }
   lazy val sampleScriptWitness: ScriptWitness = P2WPKHWitnessV0(freshXpub.key)
+
+  /**
+    * Inserts a incoming TX, and returns it with the address it was sent to
+    *
+    * This method also does some asserts on the result, to make sure what
+    * we're writing and reading matches up
+    */
+  def insertIncomingTx(daos: WalletDAOs)(implicit ec: ExecutionContext): Future[
+    (IncomingTransaction, AddressDb)] = {
+    val WalletDAOs(accountDAO, addressDAO, txDAO, _, _) = daos
+
+    /** Get a TX with outputs */
+    def getTx: Transaction =
+      TransactionGenerators.transaction
+        .suchThat(_.outputs.nonEmpty)
+        .sample
+        .getOrElse(getTx)
+
+    val account = WalletTestUtil.firstAccountDb
+
+    val address = {
+      val pub = ECPublicKey()
+      val path =
+        account.hdAccount
+          .toChain(HDChainType.External)
+          .toAddress(0)
+          .toPath
+
+      AddressDbHelper.getAddress(pub, path, RegTest)
+    }
+
+    val tx = getTx
+    val txDb = IncomingTransaction(tx,
+                                   confirmations = 3,
+                                   scriptPubKey = address.scriptPubKey)
+    for {
+      _ <- accountDAO.create(account)
+      _ <- addressDAO.create(address)
+      createdTx <- txDAO.create(txDb)
+      txAndAddr <- txDAO.withAddress(createdTx.transaction)
+    } yield
+      txAndAddr match {
+        case None =>
+          throw new org.scalatest.exceptions.TestFailedException(
+            s"Couldn't read back TX with address from DB!",
+            0)
+        case Some((foundTx, foundAddr)) =>
+          assert(foundTx.confirmations == txDb.confirmations)
+          assert(foundTx.scriptPubKey == txDb.scriptPubKey)
+          assert(foundTx.transaction == txDb.transaction)
+
+          assert(foundAddr == address)
+
+          (foundTx, foundAddr)
+      }
+
+  }
 }
