+++ wallet-test/src/test/scala/org/bitcoins/wallet/models/UTXOSpendingInfoDAOTest.scala
@@ -15,19 +15,47 @@ class UTXOSpendingInfoDAOTest extends BitcoinSWalletTest with WalletDAOFixture {
 
   it should "insert a segwit UTXO and read it" in { daos =>
     val utxoDAO = daos.utxoDAO
-    val utxo = WalletTestUtil.sampleSegwitUtxo
 
     for {
-      created <- utxoDAO.create(utxo)
+      (tx, _) <- WalletTestUtil.insertIncomingTx(daos)
+      created <- {
+        val utxo = {
+          val outpoint =
+            TransactionOutPoint(WalletTestUtil.sampleTxid,
+                                WalletTestUtil.sampleVout)
+          val output = TransactionOutput(1.bitcoin, WalletTestUtil.sampleSPK)
+          val scriptWitness = WalletTestUtil.sampleScriptWitness
+          val privkeyPath = WalletTestUtil.sampleSegwitPath
+          NativeV0UTXOSpendingInfoDb(outPoint = outpoint,
+                                     output = output,
+                                     privKeyPath = privkeyPath,
+                                     scriptWitness = scriptWitness,
+                                     incomingTxId = tx.id.get)
+        }
+        utxoDAO.create(utxo)
+      }
       read <- utxoDAO.read(created.id.get)
     } yield assert(read.contains(created))
   }
 
   it should "insert a legacy UTXO and read it" in { daos =>
     val utxoDAO = daos.utxoDAO
-    val utxo = WalletTestUtil.sampleLegacyUtxo
     for {
-      created <- utxoDAO.create(utxo)
+      (tx, _) <- WalletTestUtil.insertIncomingTx(daos)
+      created <- {
+        val utxo = {
+          val outpoint =
+            TransactionOutPoint(WalletTestUtil.sampleTxid,
+                                WalletTestUtil.sampleVout)
+          val output = TransactionOutput(1.bitcoin, WalletTestUtil.sampleSPK)
+          val privKeyPath = WalletTestUtil.sampleLegacyPath
+          LegacyUTXOSpendingInfoDb(outPoint = outpoint,
+                                   output = output,
+                                   privKeyPath = privKeyPath,
+                                   incomingTxId = tx.id.get)
+        }
+        utxoDAO.create(utxo)
+      }
       read <- utxoDAO.read(created.id.get)
     } yield assert(read.contains(created))
   }
